test: put the Docker acceptance bootstrap behind an `acc` build tag
testcontainers-go/modules/compose is imported only by provider_test.go's
TestMain, but nothing gated that import, so the Docker/containerd/
compose-go tree was compiled into the test binary for every `go test`
run -- including plain unit runs that never start a container.

Measured on this package with `go list -deps -test`:

  before   138 docker/containerd/compose-go packages
  after      0 (untagged); the 138 move behind -tags=acc

The bootstrap moves to provider_docker_acc_test.go (//go:build acc) with
a stub in provider_docker_stub_test.go (//go:build !acc) that explains
how to enable it. TestMain stays untagged, so target dispatch and unit
runs are unchanged.

Taskfile: test:acc gains -tags=acc. test:acc:hardware deliberately does
not -- it talks to an existing controller, so leaving the tag off keeps
the Docker tree out of that build too. vet gains a second tagged pass so
the acc file cannot rot silently.

One thing this does NOT do, since #159 raised it specifically: it does
not move the dependency out of go.mod's main require block. `go mod tidy`
evaluates all build configurations, so a tagged test import is still a
direct requirement -- I verified this rather than assuming it. Demoting
it for real would need the bootstrap in a separate module, which seems
disproportionate. The compile-time cost is the part that was actually
hurting, and that part is gone.
